backend/mcp-server/requirements.txt
mcp
httpx
psycopg2-binary
pgvector
sentence-transformers
